jbpm column type changed to Integer

+++ src/java/is/idega/idegaweb/egov/cases/data/GeneralCase.java
@@ -98,7 +98,7 @@ public interface GeneralCase extends IDOEntity, Case {
 	 */
 	public void setAsPrivate(boolean isPrivate);
 	
-	public abstract void setJbpmProcessInstanceId(Double pid);
+	public abstract void setJbpmProcessInstanceId(Integer pid);
 	
-	public abstract Double getJbpmProcessInstanceId();
+	public abstract Integer getJbpmProcessInstanceId();
 }
